+++ src/layouts/explainer/sections/overview/debt-chart/debt-chart.tsx
@@ -11,7 +11,8 @@ import { trillionAxisFormatter } from '../chart-helper';
 import { debtExplainerPrimary } from '../../../explainer.module.scss';
 import CustomTooltip from '../chart-components/custom-tooltip/custom-tooltip';
 import { useIsMounted } from '../../../../../utils/useIsMounted';
-import { debtEndpointUrl, deficitEndpointUrl, getOpacity, legendItems, mapBarColors, tickCountXAxis } from './debt-chart-helper';
+import { debtEndpointUrl, deficitEndpointUrl, legendItems, mapBarColors, tickCountXAxis } from './debt-chart-helper';
+import CustomBarShape from './custom-bar-shape/custom-bar-shape';
 
 const AFGDebtChart = (): ReactElement => {
   const isMounted = useIsMounted();
@@ -29,116 +30,6 @@ const AFGDebtChart = (): ReactElement => {
     'and the deficit for each year as an orange highlight against the yearly purple bars in order to represent the ' +
     'relationship between the two concepts.';
 
-  const CustomBarShape = props => {
-    const { height, width, y, x, fill, payload, dataKey, year } = props;
-    const barType = dataKey.includes('debt') ? 'debt' : 'deficit';
-    const debtVal = props[`debt${year}`];
-    const debtRemainder = debtVal % 1;
-    const debtIndexRemainder = (debtVal + (1 - debtRemainder)) % 10;
-    let countDown = payload[dataKey];
-
-    const splitWidth = (width / countDown).toFixed(1);
-    const totalWidth = splitWidth * 10;
-    const bar = (1 / 13.3) * totalWidth;
-    const gapWidth = 0.3 * bar;
-    if (width > 0) {
-      let xVal = x;
-      if (barType === 'debt') xVal = xVal + gapWidth;
-      const firstBar = bar * (1 - debtRemainder);
-      const axisGap = debtIndexRemainder % 10 === 0 ? gapWidth : 0;
-      const allBars = [];
-
-      if (barType === 'deficit') {
-        allBars.push({ x: xVal, width: firstBar });
-        xVal = xVal + firstBar + axisGap + gapWidth;
-      }
-      console.log(year, splitWidth, bar, gapWidth, barType);
-      let barCount = barType === 'debt' ? 0 : debtIndexRemainder;
-      if (barType !== 'debt') countDown = countDown - (1 - debtRemainder);
-      while (countDown >= 1) {
-        countDown = countDown - 1;
-        if (barType === 'debt') barCount = barCount + 1;
-        const debtAxisGap = barCount !== 1 && barCount % 10 === 0 ? gapWidth : 0;
-        const deficitAxisGap = barCount % 10 === 0 ? gapWidth : 0;
-        const axisGap = barType === 'debt' ? debtAxisGap : deficitAxisGap;
-        if (barType === 'deficit') barCount = barCount + 1;
-        allBars.push({ x: xVal, width: bar });
-        xVal = xVal + bar + gapWidth + axisGap;
-      }
-      allBars.push({ x: xVal, width: bar * countDown });
-
-      return (
-        <>
-          {allBars.map(val => (
-            <rect
-              {...props}
-              y={y}
-              height={height}
-              width={val.width}
-              x={val.x}
-              strokeWidth={0}
-              fill={fill}
-              fillOpacity={1}
-              opacity={getOpacity(focusedYear, year)}
-            />
-          ))}
-        </>
-      );
-    }
-  };
-
-  // const CustomDeficitBarShape = props => {
-  //   const { height, width, y, fill, payload, x, dataKey, year } = props;
-  //   const debtVal = props[`debt${year}`];
-  //   const debtRemainder = debtVal % 1;
-  //   let countDown = payload[dataKey];
-  //
-  //   const splitWidth = (width / countDown).toFixed(1);
-  //   const totalWidth = splitWidth * 10;
-  //   const bar = (1 / 13.3) * totalWidth;
-  //   const gapWidth = 0.3 * bar;
-  //   if (width > 0) {
-  //     let xVal = x;
-  //     // const barWidth = Number(((width - gap * 2) / debtVal - gap).toFixed(1));
-  //     const allBars = [];
-  //     const firstBar = bar * (1 - debtRemainder);
-  //     const debtIndexRemainder = (debtVal + (1 - debtRemainder)) % 10;
-  //     allBars.push({ x: xVal, width: firstBar });
-  //     let barCount = debtIndexRemainder;
-  //     const axisGap = barCount % 10 === 0 ? gapWidth : 0;
-  //
-  //     xVal = xVal + firstBar + gapWidth + axisGap;
-  //
-  //     countDown = countDown - (1 - debtRemainder);
-  //     while (countDown >= 1) {
-  //       countDown = countDown - 1;
-  //       const axisGap = barCount % 10 === 0 ? gapWidth : 0;
-  //       barCount = barCount + 1;
-  //       allBars.push({ x: xVal, width: bar });
-  //       xVal = xVal + bar + gapWidth + axisGap;
-  //     }
-  //
-  //     allBars.push({ x: xVal, width: bar * countDown });
-  //     return (
-  //       <>
-  //         {allBars.map(val => (
-  //           <rect
-  //             {...props}
-  //             y={y}
-  //             height={height}
-  //             width={val.width}
-  //             x={val.x}
-  //             strokeWidth={0}
-  //             fill={fill}
-  //             fillOpacity={1}
-  //             opacity={getOpacity(focusedYear, year)}
-  //           />
-  //         ))}
-  //       </>
-  //     );
-  //   }
-  // };
-
   const generateBar = sortedData => {
     return sortedData.map(yearlyData => {
       const dataYear = yearlyData.year;
@@ -146,7 +37,6 @@ const AFGDebtChart = (): ReactElement => {
       const chartData = Object.keys(yearlyData).filter(propName => {
         return propName !== 'year' && propName !== 'tooltip';
       });
-      const length = chartData.length - 1;
       return chartData.map((valueName, index) => {
         const barName = valueName === 'debt' ? `Debt` : valueName === 'deficit' ? `Deficit` : '';
         return (
